Add workaround for missing tera filter
This is required because rocket uses an old version of tera which doesn't have some of the features used in this project.
At the time of writing this, there's already a commit bumping the tera dependency's version, but it's not included in any library releases yet.
Once it is, though, this code can probably be deleted.